requirements.txt
requests
python-dotenv
openai
urllib3